================ config ================
Standalone TeX files now import the array package

================ doc ================

started backend.tex for back-end construction.  Also some other edits.

================ lua ================

Added to Lua interpreter an embedding/projection pair for a *closure*,
which could be either a Lua function or a Lua table with an `apply'
value.

================ src ================

Added some documentation to backplane.nw, in particular, started to
document properties that will be important in our back ends.

To driver2.nw, added Lua function Driver.compilec, which can accept a
closure as the thing to optimize.  Once the dust settles, this trick
should be used to simplify and regularize qc--.lua.

Added some speculation to expander.nw.

In luautil.nw, changed environment variable QC--LUAPATH to QCMMLUAPATH
so that it can actually be set by the shell.

Changed the record-field name Space.aggsize to Space.widths, as the
idea of the ``widths'' supported by a space seemed more evocative.

Added Space.checked, which checks some simple invariants on spaces.

In main.nw, substantial restructuring and renaming of the construction
of the Lua library, intended to make it a bit clearer which parts do
what.

---> Changed the meaning of command-line options, so that any argument
     containing an = sign is deemed to be a Lua chunk to be evaluated.
     This change means that it is no longer necessary to put a -e in
     front of such an option.  I've altered the man page accordingly.

Lots of restructuring of qc--.lua, but the job is only begun:

  Added function Util.funname for discovering the name of a Lua value.

  Broke up qc--.lua into pieces with some doco.

  Started adding a Toy target.

  Found Main.Util harmful (too much nesting for people), so created
  CommandLineOption for command-line options, and what was Main.Util
  is now Main.  Structure of Lua still needs more careful thought,
  however.

Added a very simple dtest to the mkfile for those of us lucky enough
to have test2.c--

Added support for ``standard'' spaces to Space.Standard module. In
flux.

Added steps toward an x86 back end in x86.nw

================ test ================

cleaned up the mkfile some, and changed `Main.' to
`CommandLineOption.' in many places.

The test suite is not clean, but I don't think it's my fault.
I'll ask Christian to have a look on return.