A Lua library can now provide strings that will be executed once the interpreter is started.  By using these strings, we will clean up the provision of initialization code that is easier to write in Lua than in Caml.  A single example of such initialization appears in the Expander module.
Implemented a generic code expander, designed to be easily specialized
for a particular target platform.

Made the test suite substantially quieter than before.  Also set up
the test suite so that its output is suitable for piping to sh (or
cutting and pasting with the mouse).  To get the old, noisy behavior,
`mk VERBOSE=-v test'.

Passes most tests, but there are problems with 6 tests as follows:

  mk VERBOSE=-v exp-000.test # (FAILED: stdout does not match)
  # tested exp-000
  mk VERBOSE=-v exp-001.test # (FAILED: stdout does not match)
  # tested test-000
  # tested test-010
  # tested test-020
  # tested test-030
  Error in test suite: cannot compare 'test-037.2' with stderr
  mk VERBOSE=-v test-038.test # (FAILED: stderr does not match)
  mk VERBOSE=-v test-039.test # (FAILED: stderr does not match)
  # tested test-040
  mk VERBOSE=-v test-042.test # (FAILED: stderr does not match)

NR