Tested call, cut to, return, and shuffle. The new expander code seems to work correctly. There is a problem with continuations not fetching from memory. John D knows about this problem.
I've also extended the expander and postexpander to deal with sums and
differences of constants, and I've added Rtlutil.add, which
automatically reassociates integer addition to put constants on the
right.  It passes all tests, though I had to change some because
arithmetic is reassociated.   It could probably be even more
aggressive, maintaining a normal form, but I have not attempted to
achieve this.

I've changed *many* clients to use Rtlutil.add.  Future clients that
need to add should use Rtlutil.add.

I've added Rtlutil.fetch and Rtlutil.store.

I've observed that operations on Bits.bits need to be refactored from
Rtleval2 into either Bitops or possibly a new module Bits.Ops.
I've started the job with Bits.add, which is probably not in the right
place.


NR