Exported interface for specifying calling convention automata out to Lua; automata may be registered with the compiler at runtime.  Switched over the compiler's method of looking up a calling convention to only look up calling convention automata registered with the new mechanism.  (The new Lua specifications for Alpha/Mips/x86 backends are found in alphacall.nw/mipscall.nw/x86call.nw and inserted into qc--.lua by the mkfile.)
Added new fields to Target.t record type to allow for above changes; all
backends must provide suitable values for these fields (see documentation
in target.nw).

Old code for automata manipulation is still present and may be reverted to
by uncommenting one line in ast2ir.nw, should that be necessary.  There is
quite a bit of dead code that may be deleted if we decide that the new code
is okay (it currently passes the test suite, passes testcc on Alpha & x86 but
MIPS still won't run because of a rounding mode-related bug)...