Overflow handler demo.
Here is the high-level overview of what happens when a stack limit check
fails (in clientbuildsclosure).

1) The function checks the stack limit and realizes it has insufficient
stack space.

2) It cuts to the overflow handler, passing a recovery continuation, the
size of the closure required to reinvoke the function, and a function that
invokes the closure (the closure-invocation function).

3) The overflow handler (running on its own stack) allocates a closure and
a
continuation that invokes the closure on a new stack; it cuts to the
recovery continuation.

4) The recovery continuation fills in the closure with a pointer to the
function, all its arguments, and a return continuation.  It then cuts to the
closure-invocation continuation, invoking the function on a new stack.

5) The function runs.

6) The function returns to the closure-invocation function.

7) The closure-invocation function cuts to the return continuation on the
original stack with the return values.

8) The return continuation returns those values to the calling function.

In runtimebuildsclosure, the overflow handler uses the runtime system to
build the closure instead.