Fixed this BUG: > we don't allocate memory in the invokeclosure frame > for the overflow return values, so if the overflow block is big enough, > it will overwrite invokeclosure local variables.  Right now, more than > two overflow return values will trash a local variable in invokeclosure.
To solve this, I added an %ebp base pointer to address local variables,
so the %esp stack pointer can float and allocate variable-sized stack space
at runtime.

Much more documentation now.  Diagrams of stack layout and everything.

I also discovered a more serious bug.  Say f overflows.  When invokeclosure
returns directly to f's caller, it doesn't restore any callee-save registers
that were saved in f's frame before it cut to the overflow handler.
At first, I thought we might be able to fish it out of the Cmm_Word* regs[]
field in the Cmm_Activation struct for f, but that field doesn't get filled
in for the youngest activation on the stack.  Trouble....