Fixed the bug: > Say f overflows.  When invokeclosure > returns directly to f's caller, it doesn't restore any callee-save registers > that were saved in f's frame before it cut to the overflow handler.
We give invokeclosure a 'body' continuation for the body of f.  Then, it
can give this to the runtime system and ask it where callee-save registers
in f are stored.  This turns out to be surprisingly painless with Norman's
help.

I have checked that callee-save registers are restored out of the same
memory locations that they are saved to.

I haven't been able to test that callee-save registers are actually preserved;
qc-- by default doesn't want to allocate my variables in callee-save registers.
With backend.ralloc=Ralloc.color it will, but then my code and the runtime system
crash, depending on which files are compiled with which register coloring algorithm.